api-service/requirements.txt
Flask
gunicorn
flask-cors
mysql-connector-python
pytz
requests